Added `executeInWorker` method in EventObserver interface
This method is an hint to indicate if the notification to the observer
should be executed in a worker separated from the default event loop of the broker.

As it is just a hint, an EventBroker may ignore it if it has no worker capabilities,
or for any other reason.